week11/requirements.txt
langchain
langchain-community
langchain-ollama
networkx
pyvis
pydantic
python-dotenv
streamlit
